+++ plwordnetMobileService/src/main/java/com/pwr/bzapps/plwordnetmobile/service/database/repository/grammar/WordRepository.java
@@ -7,14 +7,14 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface WordRepository extends CrudRepository<WordEntity, Integer> {
+public interface WordRepository extends CrudRepository<WordEntity, Long> {
     @Query(value = "SELECT CONCAT(w.id,',','\"',REPLACE(w.word,'\"','####'),'\"') FROM word w", nativeQuery = true)
     public List<String> findAllAndParseString();
 
     @Query(value = "SELECT CONCAT(w.id,',','\"',REPLACE(w.word,'\"','####'),'\"') FROM word w" +
             " WHERE w.id>=:begin AND w.id<:end", nativeQuery = true)
-    public List<String> findAllAndParseStringBatch(@Param("begin") Integer begin, @Param("end") Integer end);
+    public List<String> findAllAndParseStringBatch(@Param("begin") Long begin, @Param("end") Long end);
 
     @Query(value = "SELECT MAX(id) FROM word", nativeQuery = true)
-    public Integer getMaxIndex();
+    public Long getMaxIndex();
 }
