+++ plwordnetMobileService/src/main/java/com/pwr/bzapps/plwordnetmobile/service/database/repository/relation/RelationTypeAllowedPartOfSpeechRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface RelationTypeAllowedPartOfSpeechRepository extends CrudRepository<RelationTypeAllowedPartOfSpeechEntity, Integer> {
+public interface RelationTypeAllowedPartOfSpeechRepository extends CrudRepository<RelationTypeAllowedPartOfSpeechEntity, Long> {
     @Query(value = "SELECT CONCAT(rtapos.relation_type_id,',',rtapos.part_of_speech_id) FROM relation_type_allowed_parts_of_speech rtapos", nativeQuery = true)
     public List<String> findAllAndParseString();
 }
